Decouple admin-passthru from NVMe Wrapper to make framework modular.
